# apps/api/openapi.yaml
openapi: 3.0.3
info:
  title: RentIQ API — IAM + Organizations (Phase 2)
  description: |
    Authentication, renter identity, and organization management for the RentIQ modular
    monolith. All endpoints live under the `/api` global prefix (`servers[0]` below already
    includes it; controllers map `v1/...`).

    Error envelope (Phase 1): `{ "correlationId", "code", "message", "timestamp" }` where
    `message` is localized per ADR-006 (JWT `locale` claim → org `defaultLocale` → `uk`).
    See `specs/005-iam-organizations/contracts/api.md` for the authoritative error-code table.
  version: 0.2.0
servers:
  - url: /api
...
paths:
...
  /v1/organizations:
    post:
      tags: [Organizations]
      summary: Create an organization + bootstrap ORG_ADMIN (SUPER_ADMIN only)
      security:
        - Bearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateOrganizationRequest' }
      responses:
        '201':
          description: Organization + bootstrap admin
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CreateOrganizationResponse' }
        '400':
          description: VALIDATION_ERROR
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '403':
          description: FORBIDDEN
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '409':
          description: SLUG_TAKEN
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /v1/organizations/{id}/branding:
    patch:
      tags: [Organizations]
      summary: Update organization branding (SUPER_ADMIN only)
      security:
        - Bearer: []
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateBrandingRequest' }
      responses:
        '200':
          description: Updated organization branding
          content:
            application/json:
              schema:
                type: object
                properties:
                  organization:
                    type: object
                    properties:
                      id: { type: string, format: uuid }
                      status: { type: string, enum: [ACTIVE, SUSPENDED] }
                      branding: { $ref: '#/components/schemas/Branding' }
        '400':
          description: VALIDATION_ERROR
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '403':
          description: FORBIDDEN | ORG_SUSPENDED
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '404':
          description: NOT_FOUND
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /v1/organizations/{id}/maintenance-window:
    post:
      tags: [Organizations]
      summary: Set the org maintenance window
      description: |
        ORG_ADMIN (own org) or SUPER_ADMIN (any org) per roadmap. Actual downtime
        enforcement is owned by later capabilities; this endpoint only configures the window.
      security:
        - Bearer: []
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MaintenanceWindowRequest' }
      responses:
        '200':
          description: Updated organization with maintenance window
          content:
            application/json:
              schema:
                type: object
                properties:
                  organization:
                    type: object
                    properties:
                      id: { type: string, format: uuid }
                      maintenanceWindow: { $ref: '#/components/schemas/MaintenanceWindow' }
        '400':
          description: VALIDATION_ERROR
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '403':
          description: FORBIDDEN | ORG_SUSPENDED
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '404':
          description: NOT_FOUND
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
components:
  securitySchemes:
    Bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'Admin or renter JWT (RS256). Admin endpoints reject `type: renter` tokens and vice versa (FORBIDDEN).'
    BotSecret:
      type: http
      scheme: bearer
      description: 'Bot secret for the telegram exchange — `Authorization: Bot <secret>`.'
  schemas:
    ErrorResponse:
      type: object
      required: [correlationId, code, message, timestamp]
      properties:
        correlationId: { type: string, format: uuid }
        code: { type: string, description: 'Error code, e.g. INVALID_CREDENTIALS, CONSENT_REQUIRED' }
        message: { type: string, description: 'Localized message (ADR-006 chain: JWT locale → org defaultLocale → uk)' }
        timestamp: { type: string, format: date-time }
...
    CreateOrganizationRequest:
      type: object
      required: [name, slug, adminEmail, adminPassword, telegramBotSecret]
      properties:
        name: { type: string, minLength: 1, maxLength: 255 }
        slug: { type: string, pattern: '^[a-z0-9-]{1,63}$', description: 'Unique, lowercase' }
        adminEmail: { type: string, format: email }
        adminPassword: { type: string, format: password, minLength: 8 }
        defaultLocale: { type: string, enum: [uk, en], default: uk }
        supportedLocales: { type: array, items: { type: string, enum: [uk, en] }, default: ['uk', 'en'] }
        businessName: { type: string, description: 'Defaults to name' }
        telegramBotSecret: { type: string, description: 'Stored hashed (bcrypt); never returned' }
    CreateOrganizationResponse:
      type: object
      required: [organization, bootstrapAdmin]
      properties:
        organization:
          type: object
          required: [id, name, slug, status, branding]
          properties:
            id: { type: string, format: uuid }
            name: { type: string }
            slug: { type: string }
            status: { type: string, enum: [ACTIVE, SUSPENDED] }
            branding: { $ref: '#/components/schemas/Branding' }
        bootstrapAdmin:
          type: object
          required: [id, orgId, role, email]
          properties:
            id: { type: string, format: uuid }
            orgId: { type: string, format: uuid }
            role: { type: string, enum: [ORG_ADMIN] }
            email: { type: string, format: email }
    UpdateBrandingRequest:
      type: object
      description: All fields optional — partial update
      properties:
        logoUrl: { type: string, format: uri, nullable: true, description: 'null clears' }
        primaryColor: { type: string, pattern: '^#[0-9a-fA-F]{6}$' }
        businessName: { type: string, minLength: 1, maxLength: 255 }
        supportedLocales: { type: array, items: { type: string, enum: [uk, en] }, minItems: 1 }
        defaultLocale: { type: string, enum: [uk, en], description: 'Must be in supportedLocales' }
    Branding:
      type: object
      required: [logoUrl, primaryColor, businessName, supportedLocales, defaultLocale]
      properties:
        logoUrl: { type: string, nullable: true }
        primaryColor: { type: string, nullable: true, pattern: '^#[0-9a-fA-F]{6}$' }
        businessName: { type: string }
        supportedLocales: { type: array, items: { type: string, enum: [uk, en] } }
        defaultLocale: { type: string, enum: [uk, en] }
    MaintenanceWindowRequest:
      type: object
      required: [startTime, endTime]
      properties:
        startTime: { type: string, pattern: '^([01]\\d|2[0-3]):[0-5]\\d$', example: '02:00' }
        endTime: { type: string, pattern: '^([01]\\d|2[0-3]):[0-5]\\d$', description: 'Must be after startTime', example: '04:00' }
        timezone: { type: string, default: 'Europe/Kyiv', description: 'IANA tz name' }
    MaintenanceWindow:
      type: object
      required: [startTime, endTime, timezone]
      properties:
        startTime: { type: string }
        endTime: { type: string }
        timezone: { type: string }